test(mockElement): move into mock directory
